+++ packager/media/formats/mp2t/ts_section_pes.cc
@@ -195,7 +195,7 @@ bool TsSectionPes::ParseInternal(const uint8_t* raw_pes, int raw_pes_size) {
   RCHECK(packet_start_code_prefix == kPesStartCode);
   DVLOG(LOG_LEVEL_PES) << "stream_id=" << std::hex << stream_id << std::dec;
   if (pes_packet_length == 0)
-    pes_packet_length = bit_reader.bits_available() / 8;
+    pes_packet_length = static_cast<int>(bit_reader.bits_available()) / 8;
 
   // Ignore the PES for unknown stream IDs.
   // See ITU H.222 Table 2-22 "Stream_id assignments"
@@ -234,7 +234,7 @@ bool TsSectionPes::ParseInternal(const uint8_t* raw_pes, int raw_pes_size) {
   RCHECK(bit_reader.ReadBits(1, &pes_crc_flag));
   RCHECK(bit_reader.ReadBits(1, &pes_extension_flag));
   RCHECK(bit_reader.ReadBits(8, &pes_header_data_length));
-  int pes_header_start_size = bit_reader.bits_available() / 8;
+  int pes_header_start_size = static_cast<int>(bit_reader.bits_available()) / 8;
 
   // Compute the size and the offset of the ES payload.
   // "6" for the 6 bytes read before and including |pes_packet_length|.
@@ -287,9 +287,11 @@ bool TsSectionPes::ParseInternal(const uint8_t* raw_pes, int raw_pes_size) {
   }
 
   // Discard the rest of the PES packet header.
-  DCHECK_EQ(bit_reader.bits_available() % 8, 0);
-  int pes_header_remaining_size = pes_header_data_length -
-      (pes_header_start_size - bit_reader.bits_available() / 8);
+  DCHECK_EQ(bit_reader.bits_available() % 8, 0u);
+  int pes_header_remaining_size =
+      pes_header_data_length -
+      (pes_header_start_size -
+       static_cast<int>(bit_reader.bits_available()) / 8);
   RCHECK(pes_header_remaining_size >= 0);
 
   // Read the PES packet.
